+++ minimal-examples/crypto/minimal-crypto-x509/main.c
@@ -126,7 +126,7 @@ int main(int argc, const char **argv)
 		}
 
 		if ((p = lws_cmdline_option(argc, argv, "--alg")))
-			lws_jwk_strdup_meta(&jwk, JWK_META_ALG, p, strlen(p));
+			lws_jwk_strdup_meta(&jwk, JWK_META_ALG, p, (int)strlen(p));
 
 		lwsl_info("JWK version of trusted cert:\n");
 		lws_jwk_dump(&jwk);
@@ -143,7 +143,7 @@ int main(int argc, const char **argv)
 	lwsl_info("JWK version of cert:\n");
 
 	if ((p = lws_cmdline_option(argc, argv, "--alg")))
-		lws_jwk_strdup_meta(&jwk, JWK_META_ALG, p, strlen(p));
+		lws_jwk_strdup_meta(&jwk, JWK_META_ALG, p, (int)strlen(p));
 
 	lws_jwk_dump(&jwk);
 	/* only print public if he doesn't provide private */
@@ -172,7 +172,7 @@ int main(int argc, const char **argv)
 		}
 
 		if ((p = lws_cmdline_option(argc, argv, "--alg")))
-			lws_jwk_strdup_meta(&jwk, JWK_META_ALG, p, strlen(p));
+			lws_jwk_strdup_meta(&jwk, JWK_META_ALG, p, (int)strlen(p));
 
 		lwsl_info("JWK version of cert + privkey:\n");
 		lws_jwk_dump(&jwk);
